move hash function outside of user and into it's own file to work around a svelte issue
